back_end/auth-master/requirements.txt
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.7.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
Pillow==7.1.1
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
psycopg2-binary==2.8.6
drf-nested-routers==0.91
drf-yasg==1.20.0
django-crispy-forms==1.11.2
celery==4.4.1
redis==3.4.1